requirements.txt
pyTelegramBotAPI
requests
google-generativeai